requirements.txt
requests
pymongo
edgartools
xmltodict